Rework WP commands
- Move Monitor_ReadMemory, Monitor_WriteMemory, Monitor_EraseMemory, Monitor_QueryConfiguration, Monitor_QueryConfiguration, Monitor_UpdateConfiguration, Monitor_CheckMemory, Monitor_MemoryMap, and Monitor_FlashSectorMap implementations to a common location.
- Remove unnecessary flags check on several commands.
- Remove weak attribute on several WP commands implementation.
