+++ targets/os/win32/nanoCLR/Various.cpp
@@ -627,3 +627,10 @@ size_t CPU_GetUncachableAddress( size_t address )
 }
 
 ///////////////////////////////////////////////////////////////
+
+int nanoBooter_GetTargetInfo(TargetInfo *targetInfo)
+{
+	return 0;
+}
+
+///////////////////////////////////////////////////////////////
